fix: refactoring and tests
